requirements.txt
requests 
fire
python-dotenv
replicate
ultralytics 
opencv-python
selenium 
webdriver-manager
streamlit